+++ api/src/compiler/compiler.ts
@@ -1,7 +1,64 @@
-import { ABI, AutomaticContractVerificationReq, Compile, CompilerContracts, CompilerResult, Contracts, SolcContracts, Target, VerifyContract } from "../types";
+import { ABI, AutomaticContractVerificationReq, Target } from "../types";
 import { ensure } from "../utils";
 
 const solc = require('solc');
+
+export interface Contracts {
+  [contractFilename: string]: string
+}
+
+interface CompilerContracts {
+  [filename: string]: {
+    content: string;
+  }
+}
+
+interface SolcContracts {
+  language: 'Solidity';
+  sources: CompilerContracts;
+  settings: {
+    optimizer: {
+      runs: number;
+      enabled: boolean;
+    };
+    evmVersion?: Target;
+    outputSelection: {
+      "*": {
+        '*': string[];
+      };
+    };
+  };
+}
+
+interface CompilerResult {
+  errors?: {
+    type: string;
+    formattedMessage: string;
+  }[];
+  contracts?: {
+    [filename: string]: {
+      [name: string]: any;
+    };
+  };
+  sources: {
+    [filename: string]: {
+      id: number;
+    };
+  };
+}
+
+
+interface Bytecode {
+  core: string;
+  endMetadata: string;
+  startMetadata: string;
+}
+
+export interface Compile {
+  abi: ABI;
+  fullAbi: ABI;
+  fullBytecode: string;
+}
 
 const toCompilerContracts = (contracts: Contracts): CompilerContracts => 
   Object.keys(contracts)
@@ -29,12 +86,18 @@ const prepareSolcData = (contracts: Contracts, target: Target, enabled: boolean,
   }
 })
 
-const preprocess = (fullBytecode: string): string => {
+const preprocess = (fullBytecode: string): Bytecode => {
   const start = fullBytecode.indexOf('6080604052');
   const end = fullBytecode.indexOf('a265627a7a72315820') !== -1
     ? fullBytecode.indexOf('a265627a7a72315820')
     : fullBytecode.indexOf('a264697066735822')
-  return fullBytecode.slice(start, end);;
+
+  const context = fullBytecode.slice(start, end);
+  return {
+    startMetadata: fullBytecode.slice(0, start),
+    core: context,
+    endMetadata: fullBytecode.slice(end, fullBytecode.length)
+  };
 }
 
 const loadCompiler = async (version: string): Promise<any> => (
@@ -79,12 +142,17 @@ const compileContracts = async (name: string, filename: string, source: string, 
   }
 }
 
+interface VerifyContract {
+  abi: ABI;
+  fullAbi: ABI;
+}
 export const verifyContract = async (deployedBytecode: string, {name, filename, source, compilerVersion, target, optimization, runs}: AutomaticContractVerificationReq): Promise<VerifyContract> => {
-  const {abi, fullAbi, fullBytecode} = await compileContracts(name, filename, source, compilerVersion, target, optimization, runs);
-  const parsedBytecode = preprocess(fullBytecode);
-  const rpcBytecode = preprocess(deployedBytecode);
+  const {abi, fullAbi, fullBytecode} = await compileContracts(name, filename, source, compilerVersion, target, optimization === "true", runs);
+  const {core: parsedBytecode} = preprocess(fullBytecode);
+  const {core: rpcBytecode} = preprocess(deployedBytecode);
+  ensure(parsedBytecode === rpcBytecode, "Compiled bytecode is not the same as deployed one", 404);
 
-  ensure(parsedBytecode === rpcBytecode, "Compiled bytecode is not the same as deployed one");
+  // return compiledItems.reduce((prev, item) => [...prev, ...item.abi], [])
   return {
     abi,
     fullAbi
